requirement.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
build==1.2.2.post1
cachecontrol==0.14.2
certifi==2025.4.26
charset-normalizer==3.4.2
cleo==2.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
django==4.2.21
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
fastjsonschema==2.21.1
filelock==3.16.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
packaging==25.0
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psycopg2==2.9.9
ptyprocess==0.7.0
PyJWT==2.9.0
pyproject-hooks==1.2.0
python-decouple==3.8
pywin32-ctypes==0.2.3
rapidfuzz==3.9.7
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
trove-classifiers==2025.5.9.12
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
virtualenv==20.31.2
zipp==3.20.2
